+++ packages/create-app/templates/js-vue/src/types/message.vue
@@ -1,10 +1,10 @@
+<script>
+const click = (entity) => (entity.isUpperCase = !entity.isUpperCase)
+</script>
+
 <template>
   <span @click="click"
     >Hello
     {{ entity.isUpperCase ? entity.who.toUpperCase() : entity.who }}</span
   >
 </template>
-
-<script>
-const click = (entity) => (entity.isUpperCase = !entity.isUpperCase)
-</script>