+++ packages/create-app/templates/ts-vue/src/app.vue
@@ -1,3 +1,7 @@
+<script>
+import { Message } from "./types/message.vue"
+</script>
+
 <template>
   <h1>
     <Message v-bind="message1" />, <Message v-bind="message2" />,