+++ packages/vite-plugin-vue/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inglorious/vite-plugin-vue
 
+## 1.0.1
+
+### Patch Changes
+
+- Fix entity binding in Vue plugin
+
 ## 1.0.0
 
 ### Major Changes